Consolidated top-level interpolation functions
