+++ src/restruct_poc/beams/interpolate_QP_state.hpp
@@ -1,9 +1,6 @@
 #pragma once
 
 #include <Kokkos_Core.hpp>
-
-#include "beams.hpp"
-
 #include "src/restruct_poc/types.hpp"
 
 namespace openturbine {
@@ -118,30 +115,6 @@ struct InterpolateQPState_rprime {
         qp_rprime(j, k) = local_total[k];
     }
   }
-};
-struct InterpolateQPState {
-    Kokkos::View<Beams::ElemIndices*>::const_type elem_indices;
-    View_NxN::const_type shape_interp;
-    View_NxN::const_type shape_deriv;
-    View_N::const_type qp_jacobian;
-    View_Nx7::const_type node_u;
-    View_Nx3 qp_u;
-    View_Nx3 qp_uprime;
-    View_Nx4 qp_r;
-    View_Nx4 qp_rprime;
-
-    KOKKOS_FUNCTION
-    void operator()(Kokkos::TeamPolicy<>::member_type member) const {
-        auto i_elem = member.league_rank();
-        auto idx = elem_indices(i_elem);
-        auto first_qp = idx.qp_range.first;
-        auto first_node = idx.node_range.first;
-        auto num_nodes = idx.num_nodes;
-        Kokkos::parallel_for(Kokkos::TeamThreadRange(member, idx.num_qps), InterpolateQPState_u{first_qp, first_node, num_nodes, shape_interp, node_u, qp_u});
-        Kokkos::parallel_for(Kokkos::TeamThreadRange(member, idx.num_qps), InterpolateQPState_uprime{first_qp, first_node, num_nodes, shape_deriv, qp_jacobian, node_u, qp_uprime});
-        Kokkos::parallel_for(Kokkos::TeamThreadRange(member, idx.num_qps), InterpolateQPState_r{first_qp, first_node, num_nodes, shape_interp, node_u, qp_r});
-        Kokkos::parallel_for(Kokkos::TeamThreadRange(member, idx.num_qps), InterpolateQPState_rprime{first_qp, first_node, num_nodes, shape_deriv, qp_jacobian, node_u, qp_rprime});
-    }
 };
 
 }  // namespace openturbine
