+++ src/restruct_poc/beams/interpolate_QP_velocity.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <Kokkos_Core.hpp>
-
 #include "src/restruct_poc/types.hpp"
 
 namespace openturbine {
@@ -56,22 +55,4 @@ struct InterpolateQPVelocity_Angular {
     }
 };
 
-struct InterpolateQPVelocity {
-    Kokkos::View<Beams::ElemIndices*>::const_type elem_indices;  
-    View_NxN::const_type shape_interp;                           
-    View_Nx6::const_type node_u_dot;  
-    View_Nx3 qp_u_dot;    
-    View_Nx3 qp_omega;
-
-    KOKKOS_FUNCTION
-    void operator()(Kokkos::TeamPolicy<>::member_type member) const {
-        auto i_elem = member.league_rank();
-        auto idx = elem_indices(i_elem);
-        auto first_qp = idx.qp_range.first;
-        auto first_node = idx.node_range.first;
-        auto num_nodes = idx.num_nodes;
-        Kokkos::parallel_for(Kokkos::TeamThreadRange(member, idx.num_qps), InterpolateQPVelocity_Translational{first_qp, first_node, num_nodes, shape_interp, node_u_dot, qp_u_dot});
-        Kokkos::parallel_for(Kokkos::TeamThreadRange(member, idx.num_qps), InterpolateQPVelocity_Angular{first_qp, first_node, num_nodes, shape_interp, node_u_dot, qp_omega});
-    }
-};
 }  // namespace openturbine
