+++ src/restruct_poc/system/update_state.hpp
@@ -21,9 +21,7 @@
 #include "update_node_state.hpp"
 
 #include "src/restruct_poc/beams/beams.hpp"
-#include "src/restruct_poc/beams/interpolate_QP_acceleration.hpp"
-#include "src/restruct_poc/beams/interpolate_QP_state.hpp"
-#include "src/restruct_poc/beams/interpolate_QP_velocity.hpp"
+#include "src/restruct_poc/beams/interpolate_to_quadrature_points.hpp"
 #include "src/restruct_poc/types.hpp"
 
 namespace openturbine {
@@ -43,33 +41,21 @@ inline void UpdateState(Beams& beams, View_Nx7 Q, View_Nx6 V, View_Nx6 A) {
         }
     );
     auto range_policy = Kokkos::TeamPolicy<>(beams.num_elems, Kokkos::AUTO());
-    Kokkos::parallel_for("InterpolateQPState", range_policy, 
-        InterpolateQPState{
+    Kokkos::parallel_for("InterpolateToQuadraturePoints", range_policy,
+        InterpolateToQuadraturePoints {
             beams.elem_indices, 
             beams.shape_interp, 
             beams.shape_deriv, 
-            beams.qp_jacobian, 
-            beams.node_u, 
+            beams.qp_jacobian,
+            beams.node_u,
+            beams.node_u_dot,
+            beams.node_u_ddot,
             beams.qp_u, 
             beams.qp_u_prime, 
             beams.qp_r, 
-            beams.qp_r_prime
-        }
-    );
-    Kokkos::parallel_for("InterpolateQPVelocity", range_policy,
-        InterpolateQPVelocity{
-            beams.elem_indices,
-            beams.shape_interp,
-            beams.node_u_dot,
+            beams.qp_r_prime,
             beams.qp_u_dot,
-            beams.qp_omega
-        }
-    );
-    Kokkos::parallel_for("InterpolateQPAcceleration", range_policy,
-        InterpolateQPAcceleration {
-            beams.elem_indices,
-            beams.shape_interp,
-            beams.node_u_ddot,
+            beams.qp_omega,
             beams.qp_u_ddot,
             beams.qp_omega_dot
         }
